requirements.txt
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
black==24.1.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Deprecated==1.2.14
Django==5.0.1
django-countries==7.5.1
django-filter==23.5
django-money==3.4.1
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.29
platformdirs==4.2.0
psycopg2-binary==2.9.9
py-moneyed==3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wrapt==1.16.0
